🚚 - move Stopwatch into standalone component
